incorporated Ross and Oscar suggestion
