+++ laws/src/main/scala/cats/laws/FoldableLaws.scala
@@ -121,25 +121,6 @@ trait FoldableLaws[F[_]] extends UnorderedFoldableLaws[F] {
   def orderedConsistency[A: Eq](x: F[A], y: F[A])(implicit ev: Eq[F[A]]): IsEq[List[A]] =
     if (x === y) (F.toList(x) <-> F.toList(y))
     else List.empty[A] <-> List.empty[A]
-
-  def existsLazy[A](fa: F[A]): Boolean = {
-    var i = 0
-    F.exists(fa) { _ =>
-      i += 1
-      true
-    }
-    i == (if (F.isEmpty(fa)) 0 else 1)
-  }
-
-  def forallLazy[A](fa: F[A]): Boolean = {
-    var i = 0
-    F.forall(fa) { _ =>
-      i += 1
-      false
-    }
-    i == (if (F.isEmpty(fa)) 0 else 1)
-  }
-
 }
 
 object FoldableLaws {
