+++ laws/src/main/scala/cats/laws/discipline/FoldableTests.scala
@@ -30,8 +30,6 @@ trait FoldableTests[F[_]] extends UnorderedFoldableTests[F] {
       "foldRight is lazy" -> forAll(laws.foldRightLazy[A] _),
       "ordered constistency" -> forAll(laws.orderedConsistency[A] _),
       "exists consistent with find" -> forAll(laws.existsConsistentWithFind[A] _),
-      "exists is lazy" -> forAll(laws.existsLazy[A] _),
-      "forall is lazy" -> forAll(laws.forallLazy[A] _),
       "foldM identity" -> forAll(laws.foldMIdentity[A, B] _),
       "reduceLeftOption consistent with reduceLeftToOption" ->
         forAll(laws.reduceLeftOptionConsistentWithReduceLeftToOption[A] _),
