+++ tests/src/test/scala/cats/tests/UnorderedFoldableSuite.scala
@@ -4,8 +4,10 @@ package tests
 import org.scalacheck.Arbitrary
 import cats.instances.all._
 import cats.kernel.CommutativeMonoid
+import cats.laws.discipline.UnorderedFoldableTests
 
-sealed abstract class UnorderedFoldableSuite[F[_]](name: String)(implicit ArbFString: Arbitrary[F[String]])
+sealed abstract class UnorderedFoldableSuite[F[_]](name: String)(implicit ArbFString: Arbitrary[F[String]],
+                                                                 ArbFInt: Arbitrary[F[Int]])
     extends CatsSuite {
 
   def iterator[T](fa: F[T]): Iterator[T]
@@ -42,6 +44,7 @@ sealed abstract class UnorderedFoldableSuite[F[_]](name: String)(implicit ArbFSt
       fa.count(Function.const(true)) should ===(fa.size)
     }
   }
+  checkAll("F[Int]", UnorderedFoldableTests[F](instance).unorderedFoldable[Int, Int])
 }
 
 final class UnorderedFoldableSetSuite extends UnorderedFoldableSuite[Set]("set") {
